update upgrade to json

+++ src/main/java/net/nomUtiliser/potatoClicker/logic/data/Save.java
@@ -2,9 +2,11 @@ package net.nomUtiliser.potatoClicker.logic.data;
 
 import com.google.gson.annotations.SerializedName;
 
+import java.math.BigInteger;
+
 public class Save {
     @SerializedName("potatoes")
-    public int potatoCount;
+    public BigInteger potatoCount;
 
     public Upgrade[] upgrades;
 }
